requirements.txt
Flask
pdf2docx
python-docx
pytesseract
Pillow
